NXT-50 Refactor folders
